# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для управления арендой автомобилей (CRM).",
        "title": "Auto Hub API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/drivers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Список водителей",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу (ACTIVE, INACTIVE)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поиск по имени (ILIKE)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DriversListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Создать водителя",
                "parameters": [
                    {
                        "description": "Данные водителя",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDriverRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Driver"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/drivers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Детали водителя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID водителя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DriverDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Обновить водителя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID водителя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные водителя",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateDriverRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Driver"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "drivers"
                ],
                "summary": "Удалить водителя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID водителя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fines": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fines"
                ],
                "summary": "Список штрафов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу (UNPAID, PAID, DISPUTED)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Фильтр по машине",
                        "name": "car_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Фильтр по водителю",
                        "name": "driver_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FinesListResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Если driver_id не передан — определяется автоматически из активного договора машины",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fines"
                ],
                "summary": "Создать штраф",
                "parameters": [
                    {
                        "description": "Данные штрафа",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Fine"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/fines/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "При status=PAID проставляет paid_at. При status!=PAID сбрасывает paid_at в null.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fines"
                ],
                "summary": "Обновить штраф",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID штрафа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новый статус",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateFineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Fine"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fines"
                ],
                "summary": "Удалить штраф",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID штрафа",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Driver": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_num": {
                    "type": "string"
                },
                "passport_num": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.DriverStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.DriverStatus": {
            "type": "string",
            "enum": [
                "ACTIVE",
                "INACTIVE"
            ],
            "x-enum-varnames": [
                "DriverStatusActive",
                "DriverStatusInactive"
            ]
        },
        "domain.Fine": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "car_id": {
                    "type": "integer"
                },
                "contract_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "fine_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.FineStatus"
                }
            }
        },
        "domain.FineStatus": {
            "type": "string",
            "enum": [
                "UNPAID",
                "PAID",
                "DISPUTED"
            ],
            "x-enum-varnames": [
                "FineStatusUnpaid",
                "FineStatusPaid",
                "FineStatusDisputed"
            ]
        },
...
        "dto.ActiveContractForDriver": {
            "type": "object",
            "properties": {
                "car": {
                    "$ref": "#/definitions/dto.CarBrief"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.CarBrief": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                }
            }
        },
...
        "dto.CarForFine": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                }
            }
        },
...
        "dto.CarWithYear": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "plate_number": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ContractForFine": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ContractWithCar": {
            "type": "object",
            "properties": {
                "car": {
                    "$ref": "#/definitions/dto.CarWithYear"
                },
                "created_at": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "monthly_payment": {
                    "type": "number"
                },
                "paid_amount": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateDriverRequest": {
            "type": "object",
            "required": [
                "full_name",
                "license_num",
                "passport_num",
                "phone",
                "status"
            ],
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "license_num": {
                    "type": "string"
                },
                "passport_num": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ACTIVE",
                        "INACTIVE"
                    ]
                }
            }
        },
        "dto.CreateFineRequest": {
            "type": "object",
            "required": [
                "amount",
                "car_id",
                "description",
                "fine_date"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "car_id": {
                    "type": "integer"
                },
                "contract_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "fine_date": {
                    "type": "string"
                }
            }
        },
...
        "dto.DriverDetailResponse": {
            "type": "object",
            "properties": {
                "contracts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ContractWithCar"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_num": {
                    "type": "string"
                },
                "passport_num": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.DriverStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.DriverForFine": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.DriverWithContracts": {
            "type": "object",
            "properties": {
                "contracts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ActiveContractForDriver"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "license_num": {
                    "type": "string"
                },
                "passport_num": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.DriverStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.DriversListResponse": {
            "type": "object",
            "properties": {
                "drivers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DriverWithContracts"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.FineWithRelations": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "car": {
                    "$ref": "#/definitions/dto.CarForFine"
                },
                "car_id": {
                    "type": "integer"
                },
                "contract": {
                    "$ref": "#/definitions/dto.ContractForFine"
                },
                "contract_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "driver": {
                    "$ref": "#/definitions/dto.DriverForFine"
                },
                "driver_id": {
                    "type": "integer"
                },
                "fine_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "paid_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.FineStatus"
                }
            }
        },
        "dto.FinesListResponse": {
            "type": "object",
            "properties": {
                "fines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.FineWithRelations"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateDriverRequest": {
            "type": "object",
            "required": [
                "full_name",
                "license_num",
                "passport_num",
                "phone",
                "status"
            ],
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "license_num": {
                    "type": "string"
                },
                "passport_num": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ACTIVE",
                        "INACTIVE"
                    ]
                }
            }
        },
        "dto.UpdateFineRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "paid_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "UNPAID",
                        "PAID",
                        "DISPUTED"
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}